Move ToNXlog to file
